tabbed_choice_body added
