+++ lib/screens/home/widgets/drawer/tabbed_app_bar/tabbed_bar.dart
@@ -1,12 +1,23 @@
 import 'package:flutter/material.dart';
+import 'package:learning_compass_exp/screens/home/widgets/drawer/tabbed_app_bar/tabbed_bar_choice_body.dart';
+import 'package:learning_compass_exp/screens/home/widgets/drawer/tabbed_app_bar/tabbed_bar_choice.dart';
 
-class Choice {
-  const Choice({this.title, this.videoLink, this.body});
 
-  final String title;
-  final String videoLink;
-  final Map body;
-}
+//this will come from json
+const List choicesBodies = [
+  const ChoiceBody('title', 'body'),
+  const ChoiceBody('title', 'body'),
+  const ChoiceBody('title', 'body'),
+  const ChoiceBody('title', 'body'),
+];
+
+//this will come from json
+const List<Choice> choices = const <Choice>[
+  const Choice(title: 'OECD', body: choicesBodies ),
+  const Choice(title: 'TRENDS' ),
+  const Choice(title: 'FRAMEWORK',  ),
+  const Choice(title: 'ANALYSIS',  ),
+];
