+++ lib/screens/home/widgets/drawer/tabbed_app_bar/tabbed_bar_choice.dart
@@ -1,29 +1,14 @@
 import 'package:flutter/material.dart';
+import 'package:learning_compass_exp/screens/home/widgets/drawer/tabbed_app_bar/tabbed_bar_choice_body.dart';
 
 class Choice {
-  const Choice({this.title, this.videoLink, this.body1, this.body});
+  const Choice({this.title, this.body});
 
   final String title;
-  final String videoLink;
-  final List<ChoiceBody> body1;
-  final Map body;
+  final List<ChoiceBody> body;
+  //final Map body1;
+  //final String videoLink;
 }
-
-class ChoiceBody {
-  const ChoiceBody({this.title, this.body});
-
-  final String title;
-  final String body;
-}
-
-
-//this will come from json
-const List<Choice> choices = const <Choice>[
-  const Choice(title: 'OECD', videoLink: 'link' ),
-  const Choice(title: 'TRENDS' ),
-  const Choice(title: 'FRAMEWORK',  ),
-  const Choice(title: 'ANALYSIS',  ),
-];
 
 
 class ChoiceWidget extends StatelessWidget {
@@ -34,18 +19,19 @@ class ChoiceWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Container(
-      child: Row(
+        child: Row(
 
-        children: <Widget>[
-          Text(choice.title),
-          ListView(
             children: <Widget>[
+              Text(choice.title),
+              ListView(
+                  children: choice.body.map((ChoiceBody choicesBody) {
+                    return ChoiceBodyCard(choiceBody: choicesBody);
+                  }).toList()
+              )
+            ]
 
-              Card()
-            ],
-          )],
 
-
-      )
-   );
+        )
+    );
   }
+}
